refactor(chat-citations): move the citation preview hooks into citations/hooks
Hook modules are named `useXxx.ts` and live in a `hooks/` bucket
(naming-conventions §3.2), matching the sibling
`chat/messages/hooks/` layout.
